Day 24 slight fix (RAM is composed of longs)

+++ src/main/kotlin/io/github/clechasseur/adventofcode2021/Day24.kt
@@ -9,7 +9,7 @@ object Day24 {
     fun part1(): Long = generateSequence(99999999999999L) { it - 1L }.find { modelNumber ->
         val alu = ALU(data.lines(), modelNumber.toString())
         alu.execute()
-        alu.ram['z'] == 0
+        alu.ram['z'] == 0L
     } ?: error("No valid model number")
 
     private val inpRegex = """^inp (w|x|y|z)$""".toRegex()
@@ -20,10 +20,10 @@ object Day24 {
         private val cin = input.iterator()
 
         val ram = mutableMapOf(
-            'w' to 0,
-            'x' to 0,
-            'y' to 0,
-            'z' to 0,
+            'w' to 0L,
+            'x' to 0L,
+            'y' to 0L,
+            'z' to 0L,
         )
 
         fun execute() {
@@ -37,7 +37,7 @@ object Day24 {
             val inp = inpRegex.matchEntire(line)
             if (inp != null) {
                 require(cin.hasNext()) { "Empty input stream" }
-                ram[inp.groupValues[1].first()] = cin.nextChar().toString().toInt()
+                ram[inp.groupValues[1].first()] = cin.nextChar().toString().toLong()
             } else {
                 val cmd = cmdRegex.matchEntire(line) ?: error("Wrong program line: $line")
                 val (instruction, dest, src) = cmd.destructured
@@ -46,16 +46,16 @@ object Day24 {
                 when (instruction) {
                     "add" -> ram[dest.first()] = destValue + srcValue
                     "mul" -> ram[dest.first()] = destValue * srcValue
-                    "div" -> ram[dest.first()] = floor(destValue.toDouble() / srcValue.toDouble()).toInt()
+                    "div" -> ram[dest.first()] = floor(destValue.toDouble() / srcValue.toDouble()).toLong()
                     "mod" -> ram[dest.first()] = destValue % srcValue
-                    "eql" -> ram[dest.first()] = if (destValue == srcValue) 1 else 0
+                    "eql" -> ram[dest.first()] = if (destValue == srcValue) 1L else 0L
                     else -> error("Invalid instruction: $instruction")
                 }
             }
         }
 
-        private fun getValue(varOrVal: String): Int = if (varOrVal.length == 1 && varOrVal.first().isLetter()) {
+        private fun getValue(varOrVal: String): Long = if (varOrVal.length == 1 && varOrVal.first().isLetter()) {
             ram[varOrVal.first()] ?: error("Wrong variable: $varOrVal")
-        } else varOrVal.toInt()
+        } else varOrVal.toLong()
     }
 }
